- update of irstlm library: files larger than 4Gb can be handled in   mmap (see irstlm/src/*cpp and irstlm/src/*h) - fixed a bug in querying IRST LMs with OOVs (LanguageModelIRST.cpp) - some more checks on config file: if specified, existence of generation   and distortion files is checked (Parameter.cpp) - 0 valued entries in binary phrase tables are loaded as 1.0e-38   (PhraseDictionaryTree.cpp)
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk@1189 1f5c12ca-751b-0410-a591-d2e778427230

+++ irstlm/src/n_gram.h
@@ -47,7 +47,7 @@ class ngram{
   int  midx[MAX_NGRAM];  // ngram-tree scan pointer
   int    lev;            // ngram-tree level
   int   size;            // ngram size
-  int   freq;            // ngram frequency or integer prob
+  long long   freq;      // ngram frequency or integer prob
   int   succ;            // number of successors
   int   bow;             // back-off weight 
   int   prob;            // probability
